Refactored to change listeners obj in state and to handle listeners adds.
