+++ client/utils/handleTracklistPostResponse.js
@@ -1,33 +1,10 @@
-const state = require('./../state/state.js')
-const { addLoadRef } = require('./../state/elementRefs')
 const { attachOnZipListener } = require('./../listeners/socketListeners')
+const attachTracklistFinalListener = require('./../listeners/tracklistFinal')
 const { addListener } = require('./../listeners/index')
 
 const handleTracklistPostResponse = response => {
-  const $trackFinalContainer = addLoadRef('track-final-container')
-  const $audioPlayer = addLoadRef('audio-player')
-  const $nowPlaying = addLoadRef('now-playing')
-
   if (response.status === 202) {
-    $trackFinalContainer.addEventListener('click', e => {
-      state.selectedTrack = state.tracklist[(parseInt(e.target.dataset.tracknum, 10) - 1)]
-      if (!state.selectedTrack) return
-      state.audio.selectTrack(state.selectedTrack)
-      const $selected = addLoadRef(`track-final-${e.target.dataset.tracknum}`)
-      for (let i = 1; i <= state.tracklist.length; i++) {
-        const $track = addLoadRef(`track-final-${i}`)
-        if ($track.classList.value.includes('selected')) $track.classList.remove('selected')
-      }
-      $selected.setAttribute('class', 'track-final selected')
-
-      $nowPlaying.textContent = state.selectedTrack.trackName
-      const trackFileName = state.selectedTrack.trackName.split(' ').join('-')
-
-      const trackPath = `/download/${state.albumMetadata.videoId}/tracks/${state.socketId}/${trackFileName}.mp3`
-      $audioPlayer.pause()
-      $audioPlayer.src = trackPath
-      $audioPlayer.play()
-    })
+    addListener('tracklistFinal', attachTracklistFinalListener)
     addListener('socketOnZip', attachOnZipListener)
   }
   else (console.log('Tracklist request failed.'))
